Uploaded Excel Dashboard Project
This dashboard provides a YoY executive summary of the call data from a company's Insurance phone line (June 2021 - June 2022). The Insurance phone line had 2 call options: 1 for Authorizations, and 2 for Billing.

All charts, tables, and graphs are images, so there is no worry about messing up the filters.

Disclaimer: This is an executive summary of the call data. Discrepancies were noted, but these would be answered if:
a) The stakeholder wanted further analysis.
b) The data did not provide enough information to perform further analysis.